Set empty string as default value for backwards compatibility
Not setting the `:xcarchive` parameter (old users + users that don't
want this feature) sets the internal xcarchive file path to an empty
string so the xcarchive upload method is skipped.

Setting the `:xcarchive` parameter to a valid path to an `.xcarchive`
file directly uses it as input for the xcarchive upload method.
Setting it to `nil` causes it to load the path from the shared
`XCODEBUILD_ARCHIVE` value that's set by _gym_.